moved some stuff around, it's now possible to simulate buffs on minions
