+++ LuckParser/Models/ParseModels/Players/Minion.cs
@@ -1,5 +1,7 @@
 ﻿using LuckParser.Models.DataModels;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace LuckParser.Models.ParseModels
 {
@@ -51,6 +53,81 @@ namespace LuckParser.Models.ParseModels
         protected override void SetDamageTakenLogs(ParsedLog log)
         {
             throw new InvalidOperationException();
+        }
+
+        protected override void SetBoonDistribution(ParsedLog log)
+        {
+            List<PhaseData> phases = log.FightData.GetPhases(log);
+            BoonMap toUse = GetBoonMap(log);
+            long dur = log.FightData.FightDuration;
+            int fightDuration = (int)(dur) / 1000;
+            for (int i = 0; i < phases.Count; i++)
+            {
+                BoonDistribution.Add(new BoonDistribution());
+            }
+
+            long death = GetDeath(log, 0, dur);
+            foreach (Boon boon in TrackedBoons)
+            {
+                long boonid = boon.ID;
+                if (toUse.TryGetValue(boonid, out List<BoonLog> logs) && logs.Count != 0)
+                {
+                    if (BoonDistribution[0].ContainsKey(boonid))
+                    {
+                        continue;
+                    }
+                    BoonSimulator simulator = boon.CreateSimulator(log);
+                    simulator.Simulate(logs, dur);
+                    if (death > 0 && GetCastLogs(log, death + 5000, dur).Count == 0)
+                    {
+                        simulator.Trim(death);
+                    }
+                    else
+                    {
+                        simulator.Trim(log.FightData.ToFightSpace(LastAware));
+                    }
+                    GenerationSimulationResult generationSimulation = simulator.GenerationSimulationResult;
+                    List<BoonsGraphModel.Segment> graphSegments = new List<BoonsGraphModel.Segment>();
+                    foreach (BoonSimulationItem simul in generationSimulation.Items)
+                    {
+                        for (int i = 0; i < phases.Count; i++)
+                        {
+                            PhaseData phase = phases[i];
+                            simul.SetBoonDistributionItem(BoonDistribution[i], phase.Start, phase.End, boonid);
+                        }
+                        BoonsGraphModel.Segment segment = simul.ToSegment();
+                        if (graphSegments.Count == 0)
+                        {
+                            graphSegments.Add(new BoonsGraphModel.Segment(0, segment.Start, 0));
+                        }
+                        else if (graphSegments.Last().End != segment.Start)
+                        {
+                            graphSegments.Add(new BoonsGraphModel.Segment(graphSegments.Last().End, segment.Start, 0));
+                        }
+                        graphSegments.Add(segment);
+                    }
+                    List<AbstractBoonSimulationItem> extraSimulations = new List<AbstractBoonSimulationItem>(simulator.OverstackSimulationResult);
+                    extraSimulations.AddRange(simulator.WasteSimulationResult);
+                    foreach (AbstractBoonSimulationItem simul in extraSimulations)
+                    {
+                        for (int i = 0; i < phases.Count; i++)
+                        {
+                            PhaseData phase = phases[i];
+                            simul.SetBoonDistributionItem(BoonDistribution[i], phase.Start, phase.End, boonid);
+                        }
+                    }
+                    
+                    if (graphSegments.Count > 0)
+                    {
+                        graphSegments.Add(new BoonsGraphModel.Segment(graphSegments.Last().End, dur, 0));
+                    }
+                    else
+                    {
+                        graphSegments.Add(new BoonsGraphModel.Segment(0, dur, 0));
+                    }
+                    BoonPoints[boonid] = new BoonsGraphModel(boon, graphSegments);
+                }
+            }
         }
 
         /*protected override void setHealingLogs(ParsedLog log)
